Move index/fonts files; Update imports
