+++ application/ui/src/index.scss
@@ -2,7 +2,7 @@
  * Copyright (C) 2025 Intel Corporation
  * SPDX-License-Identifier: Apache-2.0
  */
-@import './fonts.css';
+@import './assets/fonts/fonts.css';
 
 * {
     box-sizing: border-box;
